refactor(admin): split AdminDemoController into demo state/pipeline/chat services (max-params 2/3)
AdminDemoController injected 5 deps (surreal, ingest, search, dreams,
chatRouter) and imported src/db directly (layer leak). Three-way split:
- DemoStateService (surreal) — accumulated-state counts, reset, known
  entity names. Drops the controller's src/db import.
- DemoPipelineService (ingest, search, dreams) — ingest-mention, search,
  dreams ops, and the chat tell/ask flows.
- DemoChatService (chatRouter, pipeline, state) — chat orchestration
  (knownNames → route → tell/ask).
The controller keeps HTTP plumbing (validation, PII-scope gate,
debug-trace capture, response shaping, upstream-LLM 503 classification)
with (pipeline, state, chatService), 3 deps.

Also finishes the layer-purity C-series for the demo controller.

Part of the max-params=3 god-class split program.
